pyproject.toml
[tool.poetry]
name = "rimworld-wiki-chat-bot"
version = "0.1.0"
description = ""
authors = ["Kenny <kennylink2626@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
psycopg = {extras = ["binary"], version = "^3.2.1"}
python-dotenv = "^1.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
pydantic = "^2.8.2"
markdownify = "^0.13.1"
streamlit = "^1.38.0"
openai = "^1.43.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
